// frontend/src/actions/proto/create_scenario.ts
export interface Objective {
  letter: string;
  state: ObjectiveState;
  position: Position | undefined;
}

export interface Unit {
  position: Position | undefined;
  unitKey: string;
  side: UnitSide;
  icon: string;
}

export interface ScenarioArea {
  type: string;
  coordinates: Ring[];
}

export interface Ring {
  points: Position[];
}

export interface Scenario {
  name: string;
  objectives: Objective[];
  units: Unit[];
  areas: ScenarioArea[];
}
// ...
export const Objective: MessageFns<Objective> = {
  encode(message: Objective, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.letter !== "") {
      writer.uint32(10).string(message.letter);
    }
    if (message.state !== 0) {
      writer.uint32(16).int32(message.state);
    }
    if (message.position !== undefined) {
      Position.encode(message.position, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Objective {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseObjective();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.letter = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.state = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.position = Position.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Objective {
    return {
      letter: isSet(object.letter) ? globalThis.String(object.letter) : "",
      state: isSet(object.state) ? objectiveStateFromJSON(object.state) : 0,
      position: isSet(object.position) ? Position.fromJSON(object.position) : undefined,
    };
  },

  toJSON(message: Objective): unknown {
    const obj: any = {};
    if (message.letter !== "") {
      obj.letter = message.letter;
    }
    if (message.state !== 0) {
      obj.state = objectiveStateToJSON(message.state);
    }
    if (message.position !== undefined) {
      obj.position = Position.toJSON(message.position);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Objective>, I>>(base?: I): Objective {
    return Objective.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Objective>, I>>(object: I): Objective {
    const message = createBaseObjective();
    message.letter = object.letter ?? "";
    message.state = object.state ?? 0;
    message.position = (object.position !== undefined && object.position !== null)
      ? Position.fromPartial(object.position)
      : undefined;
    return message;
  },
};

function createBaseUnit(): Unit {
  return { position: undefined, unitKey: "", side: 0, icon: "" };
}

export const Unit: MessageFns<Unit> = {
  encode(message: Unit, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.position !== undefined) {
      Position.encode(message.position, writer.uint32(10).fork()).join();
    }
    if (message.unitKey !== "") {
      writer.uint32(18).string(message.unitKey);
    }
    if (message.side !== 0) {
      writer.uint32(24).int32(message.side);
    }
    if (message.icon !== "") {
      writer.uint32(34).string(message.icon);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Unit {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUnit();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.position = Position.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.unitKey = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.side = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.icon = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Unit {
    return {
      position: isSet(object.position) ? Position.fromJSON(object.position) : undefined,
      unitKey: isSet(object.unitKey) ? globalThis.String(object.unitKey) : "",
      side: isSet(object.side) ? unitSideFromJSON(object.side) : 0,
      icon: isSet(object.icon) ? globalThis.String(object.icon) : "",
    };
  },

  toJSON(message: Unit): unknown {
    const obj: any = {};
    if (message.position !== undefined) {
      obj.position = Position.toJSON(message.position);
    }
    if (message.unitKey !== "") {
      obj.unitKey = message.unitKey;
    }
    if (message.side !== 0) {
      obj.side = unitSideToJSON(message.side);
    }
    if (message.icon !== "") {
      obj.icon = message.icon;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Unit>, I>>(base?: I): Unit {
    return Unit.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Unit>, I>>(object: I): Unit {
    const message = createBaseUnit();
    message.position = (object.position !== undefined && object.position !== null)
      ? Position.fromPartial(object.position)
      : undefined;
    message.unitKey = object.unitKey ?? "";
    message.side = object.side ?? 0;
    message.icon = object.icon ?? "";
    return message;
  },
};

function createBaseScenarioArea(): ScenarioArea {
  return { type: "", coordinates: [] };
}

export const ScenarioArea: MessageFns<ScenarioArea> = {
  encode(message: ScenarioArea, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== "") {
      writer.uint32(10).string(message.type);
    }
    for (const v of message.coordinates) {
      Ring.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ScenarioArea {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseScenarioArea();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.type = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.coordinates.push(Ring.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ScenarioArea {
    return {
      type: isSet(object.type) ? globalThis.String(object.type) : "",
      coordinates: globalThis.Array.isArray(object?.coordinates)
        ? object.coordinates.map((e: any) => Ring.fromJSON(e))
        : [],
    };
  },

  toJSON(message: ScenarioArea): unknown {
    const obj: any = {};
    if (message.type !== "") {
      obj.type = message.type;
    }
    if (message.coordinates?.length) {
      obj.coordinates = message.coordinates.map((e) => Ring.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ScenarioArea>, I>>(base?: I): ScenarioArea {
    return ScenarioArea.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ScenarioArea>, I>>(object: I): ScenarioArea {
    const message = createBaseScenarioArea();
    message.type = object.type ?? "";
    message.coordinates = object.coordinates?.map((e) => Ring.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const Scenario: MessageFns<Scenario> = {
  encode(message: Scenario, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    for (const v of message.objectives) {
      Objective.encode(v!, writer.uint32(18).fork()).join();
    }
    for (const v of message.units) {
      Unit.encode(v!, writer.uint32(26).fork()).join();
    }
    for (const v of message.areas) {
      ScenarioArea.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Scenario {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseScenario();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.objectives.push(Objective.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.units.push(Unit.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.areas.push(ScenarioArea.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Scenario {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      objectives: globalThis.Array.isArray(object?.objectives)
        ? object.objectives.map((e: any) => Objective.fromJSON(e))
        : [],
      units: globalThis.Array.isArray(object?.units) ? object.units.map((e: any) => Unit.fromJSON(e)) : [],
      areas: globalThis.Array.isArray(object?.areas) ? object.areas.map((e: any) => ScenarioArea.fromJSON(e)) : [],
    };
  },

  toJSON(message: Scenario): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.objectives?.length) {
      obj.objectives = message.objectives.map((e) => Objective.toJSON(e));
    }
    if (message.units?.length) {
      obj.units = message.units.map((e) => Unit.toJSON(e));
    }
    if (message.areas?.length) {
      obj.areas = message.areas.map((e) => ScenarioArea.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Scenario>, I>>(base?: I): Scenario {
    return Scenario.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Scenario>, I>>(object: I): Scenario {
    const message = createBaseScenario();
    message.name = object.name ?? "";
    message.objectives = object.objectives?.map((e) => Objective.fromPartial(e)) || [];
    message.units = object.units?.map((e) => Unit.fromPartial(e)) || [];
    message.areas = object.areas?.map((e) => ScenarioArea.fromPartial(e)) || [];
    return message;
  },
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}